Refactor xMultiple messages to use service.Create or service.Update methods as opposed to the internal ones.
